Skip all but one large test in strided reduce scatter nightly files
Mark all pytest.param configs as skipped (run manually) except one
large representative per nightly file:
  - t3000 strided RS: experimental_strided_128x16_8_cores_2_N_blocks_6_workers
  - tg strided RS:    strided_large_8_cores_6_workers
  - t3000 fused MM:   xlarge_4k_Nwt16_cwimb2
  - tg fused MM:      xlarge_9472_3456_5120_y7_cwimb1_rs3_fullgrid

Add wan2_2 model-specific test files with one large representative test
from each of the four nightly files:
  - test_strided_reduce_scatter_wan_t3000.py (1x8 mesh)
  - test_strided_reduce_scatter_wan_tg.py    (4x8 mesh, BH fused MM uses 12x8 grid)
