+++ tests/nightly/tg/ccl/test_strided_reduce_scatter_async.py
@@ -46,6 +46,7 @@ def _make_fabric_router_config(max_packet_payload_size_bytes):
                 chunk_width_in_mm_blocks=2,
             ),
             id="strided_toy_4x4",
+            marks=pytest.mark.skip(reason="run manually"),
         ),
         pytest.param(
             ReduceScatterTestConfig(
@@ -69,6 +70,7 @@ def _make_fabric_router_config(max_packet_payload_size_bytes):
                 chunk_width_in_mm_blocks=1,
             ),
             id="strided_4x4_2core_grid",
+            marks=pytest.mark.skip(reason="run manually"),
         ),
         pytest.param(
             ReduceScatterTestConfig(
@@ -92,6 +94,7 @@ def _make_fabric_router_config(max_packet_payload_size_bytes):
                 chunk_width_in_mm_blocks=2,
             ),
             id="strided_asymmetric_tall_narrow_8x2_blocks",
+            marks=pytest.mark.skip(reason="run manually"),
         ),
         pytest.param(
             ReduceScatterTestConfig(
@@ -115,6 +118,79 @@ def _make_fabric_router_config(max_packet_payload_size_bytes):
                 chunk_width_in_mm_blocks=2,
             ),
             id="strided_partial_last_chunk",
+            marks=pytest.mark.skip(reason="run manually"),
+        ),
+        pytest.param(
+            ReduceScatterTestConfig(
+                rs_input_shape=[2, 1, 9472, 5120],
+                dim=3,
+                layout=ttnn.TILE_LAYOUT,
+                rs_input_dtype=ttnn.bfloat16,
+                use_new=False,
+                enable_trace=False,
+                num_iters=1,
+                use_barrier=True,
+                use_persistent_buffers=True,
+                use_strided=True,
+                verify_output_shape=True,
+                verify_output_pcc=True,
+                small_random_ints=True,
+                mm_cores_y=7,
+                mm_block_ht=8,
+                mm_block_wt=8,
+                mm_N_full_block_wt=20,
+                chunk_width_in_mm_blocks=1,
+            ),
+            id="strided_large_non_div_Ht",
+            marks=pytest.mark.skip(reason="run manually"),
+        ),
+        pytest.param(
+            ReduceScatterTestConfig(
+                rs_input_shape=[8, 1, 128, 1792],
+                dim=3,
+                layout=ttnn.TILE_LAYOUT,
+                rs_input_dtype=ttnn.bfloat16,
+                use_new=False,
+                enable_trace=False,
+                num_iters=1,
+                use_barrier=True,
+                use_persistent_buffers=True,
+                use_strided=True,
+                verify_output_shape=True,
+                verify_output_pcc=True,
+                small_random_ints=True,
+                mm_cores_y=1,
+                mm_block_ht=4,
+                mm_block_wt=3,
+                mm_N_full_block_wt=9,
+                chunk_width_in_mm_blocks=1,
+            ),
+            id="strided_non_div_Wt_cross_col",
+            marks=pytest.mark.skip(reason="run manually"),
+        ),
+        pytest.param(
+            ReduceScatterTestConfig(
+                rs_input_shape=[8, 1, 224, 768],
+                dim=3,
+                layout=ttnn.TILE_LAYOUT,
+                rs_input_dtype=ttnn.bfloat16,
+                use_new=False,
+                enable_trace=False,
+                num_iters=1,
+                use_barrier=True,
+                use_persistent_buffers=True,
+                use_strided=True,
+                verify_output_shape=True,
+                verify_output_pcc=True,
+                small_random_ints=True,
+                mm_cores_y=2,
+                mm_block_ht=3,
+                mm_block_wt=1,
+                mm_N_full_block_wt=2,
+                chunk_width_in_mm_blocks=1,
+            ),
+            id="strided_non_div_Wt_and_Ht",
+            marks=pytest.mark.skip(reason="run manually"),
         ),
         pytest.param(
             ReduceScatterTestConfig(
@@ -139,75 +215,6 @@ def _make_fabric_router_config(max_packet_payload_size_bytes):
                 num_workers_per_link=6,
             ),
             id="strided_large_8_cores_6_workers",
-        ),
-        pytest.param(
-            ReduceScatterTestConfig(
-                rs_input_shape=[2, 1, 9472, 5120],
-                dim=3,
-                layout=ttnn.TILE_LAYOUT,
-                rs_input_dtype=ttnn.bfloat16,
-                use_new=False,
-                enable_trace=False,
-                num_iters=1,
-                use_barrier=True,
-                use_persistent_buffers=True,
-                use_strided=True,
-                verify_output_shape=True,
-                verify_output_pcc=True,
-                small_random_ints=True,
-                mm_cores_y=7,
-                mm_block_ht=8,
-                mm_block_wt=8,
-                mm_N_full_block_wt=20,
-                chunk_width_in_mm_blocks=1,
-            ),
-            id="strided_large_non_div_Ht",
-        ),
-        pytest.param(
-            ReduceScatterTestConfig(
-                rs_input_shape=[8, 1, 128, 1792],
-                dim=3,
-                layout=ttnn.TILE_LAYOUT,
-                rs_input_dtype=ttnn.bfloat16,
-                use_new=False,
-                enable_trace=False,
-                num_iters=1,
-                use_barrier=True,
-                use_persistent_buffers=True,
-                use_strided=True,
-                verify_output_shape=True,
-                verify_output_pcc=True,
-                small_random_ints=True,
-                mm_cores_y=1,
-                mm_block_ht=4,
-                mm_block_wt=3,
-                mm_N_full_block_wt=9,
-                chunk_width_in_mm_blocks=1,
-            ),
-            id="strided_non_div_Wt_cross_col",
-        ),
-        pytest.param(
-            ReduceScatterTestConfig(
-                rs_input_shape=[8, 1, 224, 768],
-                dim=3,
-                layout=ttnn.TILE_LAYOUT,
-                rs_input_dtype=ttnn.bfloat16,
-                use_new=False,
-                enable_trace=False,
-                num_iters=1,
-                use_barrier=True,
-                use_persistent_buffers=True,
-                use_strided=True,
-                verify_output_shape=True,
-                verify_output_pcc=True,
-                small_random_ints=True,
-                mm_cores_y=2,
-                mm_block_ht=3,
-                mm_block_wt=1,
-                mm_N_full_block_wt=2,
-                chunk_width_in_mm_blocks=1,
-            ),
-            id="strided_non_div_Wt_and_Ht",
         ),
     ],
 )
